refactor: update navigation structure and fix routing issues
- Replaced all `<a>` tags with Next.js `<Link>` components to prevent page reloads and maintain authentication state during navigation.
- Updated event URLs to use the plural form `/events/[eventSlug]/` for consistency across the application.
- Enhanced the AppSidebar to include role-based navigation and event-specific links.
- Removed deprecated event pages and streamlined the routing logic for better user experience.
- Added new pages for event code of conduct and improved event dashboard functionality.
- Documented navigation architecture and troubleshooting steps for better developer guidance.
