--- frontend/pages/event/[event-slug]/admin/reports/user/[user-id].tsx
@@ -1,10 +1,0 @@
-import React from "react";
-
-export default function AdminUserReportsPage() {
-  return (
-    <div className="p-6">
-      <h1 className="text-xl font-bold">Admin: User Reports Page</h1>
-      <p className="text-gray-600">This page is under development.</p>
-    </div>
-  );
-}
