+++ website/docs/user-guide/getting-started.md
@@ -5,6 +5,44 @@ sidebar_position: 2
 
 Welcome to the Conducky User Guide!
 
+## First Steps
+
+After logging into Conducky, you'll be directed to your appropriate starting point:
+
+- **First-time users**: Global dashboard with information about joining events
+- **Single event users**: Directly to that event's dashboard
+- **Multi-event users**: Global dashboard showing all your events
+- **SuperAdmins**: System admin dashboard for installation management
+
+## Understanding Navigation
+
+Conducky uses a three-level navigation system:
+
+1. **Global Dashboard** (`/dashboard`) - Your multi-event overview
+2. **Event Context** (`/events/[eventSlug]/`) - Event-specific functionality
+3. **System Admin** (`/admin/`) - Installation management (SuperAdmins only)
+
+See the [Navigation Guide](./navigation.md) for detailed information about how navigation works.
+
+## Your First Event
+
+### Joining an Event
+
+Most users join events through invitation links sent by event administrators. When you receive an invitation:
+
+1. Click the invitation link
+2. Log in or create an account if needed
+3. Accept the invitation to join the event with the specified role
+4. You'll be redirected to the event dashboard
+
+### Understanding Your Role
+
+Your role in each event determines what you can see and do:
+
+- **Reporter**: Submit reports, view your own reports, see event information
+- **Responder**: Handle reports, view all reports, manage report states
+- **Admin**: Full event management, team management, event settings
+
 ## Environment Variables
 
 Before running Conducky, you must set the required environment variables for both backend and frontend. See the [Developer Docs Introduction](../developer-docs/intro.md#environment-variables-standardized) for the full list and details.
